Very early ELF writing w/ assembly

+++ src/main/kotlin/coder/format/elf/header/ELFSectionHeaderFlags.kt
@@ -1,6 +1,6 @@
 package org.bread_experts_group.coder.format.elf.header
 
-enum class ELFSectionHeaderFlags(val position: Int) {
+enum class ELFSectionHeaderFlags(val position: Long) {
 	SHF_WRITE(0x1),
 	SHF_ALLOC(0x2),
 	SHF_EXECINSTR(0x4),
